requirements.txt
numpy
pandas
pillow
six
torch
torchvision
tqdm